Add ディレクトリ構成を調整、apiClientの修正
