+++ front/src/components/atoms/GoogleAuthButton.tsx
@@ -5,5 +5,5 @@ export const GoogleAuthButton = () => {
   const handleSignIn = () => {
     signIn('google')
   }
-  return <Button variant="outline" onClick={handleSignIn}>認証ボタン</Button>
+  return <Button variant="outline" onClick={handleSignIn}>Google認証ボタン</Button>
 }
