+++ front/src/lib/auth/options.ts
